secret-manager-rotator/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-secret-manager==2.5.0s
